+++ sdk/core/azure_core/src/http/policies/mod.rs
@@ -4,8 +4,10 @@
 //! HTTP pipeline policies.
 
 mod bearer_token_policy;
+mod client_request_id;
 mod telemetry;
 
 pub use bearer_token_policy::BearerTokenCredentialPolicy;
+pub use client_request_id::*;
 pub use telemetry::*;
 pub use typespec_client_core::http::policies::*;